requirements.txt
scikit-learn>=0.22
scipy>=1.3.2
numpy>=1.17.4
torch>=1.00
torchvision>=0.2.2
matplotlib>=3.0.3
tqdm>=4.40.2
efficientnet-pytorch>=0.6.3
neptune-client
tensorboard
pycocotools